move `Spans` and `NextNeededPruningSeed`
