+++ binaries/cuprated/src/rpc/request/address_book.rs
@@ -177,53 +177,3 @@ pub(crate) async fn get_bans<Z: NetworkZone>(
 
     Ok(bans)
 }
-
-/// [`AddressBookRequest::Spans`]
-pub(crate) async fn spans<Z: NetworkZone>(
-    address_book: &mut impl AddressBook<Z>,
-) -> Result<Vec<Span>, Error> {
-    let AddressBookResponse::Spans(vec) = address_book
-        .ready()
-        .await
-        .map_err(|e| anyhow!(e))?
-        .call(AddressBookRequest::Spans)
-        .await
-        .map_err(|e| anyhow!(e))?
-    else {
-        unreachable!();
-    };
-
-    // FIXME: impl this map somewhere instead of inline.
-    let vec = vec
-        .into_iter()
-        .map(|span| Span {
-            connection_id: String::from(FIELD_NOT_SUPPORTED),
-            nblocks: span.nblocks,
-            rate: span.rate,
-            remote_address: span.remote_address.to_string(),
-            size: span.size,
-            speed: span.speed,
-            start_block_height: span.start_block_height,
-        })
-        .collect();
-
-    Ok(vec)
-}
-
-/// [`AddressBookRequest::NextNeededPruningSeed`]
-pub(crate) async fn next_needed_pruning_seed<Z: NetworkZone>(
-    address_book: &mut impl AddressBook<Z>,
-) -> Result<PruningSeed, Error> {
-    let AddressBookResponse::NextNeededPruningSeed(seed) = address_book
-        .ready()
-        .await
-        .map_err(|e| anyhow!(e))?
-        .call(AddressBookRequest::NextNeededPruningSeed)
-        .await
-        .map_err(|e| anyhow!(e))?
-    else {
-        unreachable!();
-    };
-
-    Ok(seed)
-}
